@@ -7,13 +7,26 @@ import logging
 import time
 import asyncio
 from typing import List, Dict, Any, AsyncGenerator, Union, Optional, Tuple
-from .provider_interface import ProviderInterface
+from .provider_interface import ProviderInterface, QuotaGroupMap, UsageResetConfigDef
 from .gemini_auth_base import GeminiAuthBase
 from .provider_cache import ProviderCache
-from .antigravity_provider import GEMINI3_TOOL_RENAMES, GEMINI3_TOOL_RENAMES_REVERSE
+from .utilities.gemini_cli_quota_tracker import GeminiCliQuotaTracker
+from .utilities.gemini_shared_utils import (
+    env_bool,
+    env_int,
+    inline_schema_refs,
+    recursively_parse_json_strings,
+    GEMINI3_TOOL_RENAMES,
+    GEMINI3_TOOL_RENAMES_REVERSE,
+    FINISH_REASON_MAP,
+    CODE_ASSIST_ENDPOINT,
+)
+from ..transaction_logger import ProviderLogger
+from .utilities.gemini_tool_handler import GeminiToolHandler
+from .utilities.gemini_credential_manager import GeminiCredentialManager
 from ..model_definitions import ModelDefinitions
 from ..timeout_config import TimeoutConfig
-from ..utils.paths import get_logs_dir, get_cache_dir
+from ..utils.paths import get_cache_dir
 import litellm
 from litellm.exceptions import RateLimitError
 from ..error_handler import extract_retry_after_from_body
@@ -23,13 +36,6 @@ import uuid
 from datetime import datetime
 
 lib_logger = logging.getLogger("rotator_library")
-
-
-def _get_gemini_cli_logs_dir() -> Path:
-    """Get the Gemini CLI logs directory."""
-    logs_dir = get_logs_dir() / "gemini_cli_logs"
-    logs_dir.mkdir(parents=True, exist_ok=True)
-    return logs_dir
 
 
 def _get_gemini_cli_cache_dir() -> Path:
@@ -42,77 +48,11 @@ def _get_gemini3_signature_cache_file() -> Path:
     return _get_gemini_cli_cache_dir() / "gemini3_signatures.json"
 
 
-class _GeminiCliFileLogger:
-    """A simple file logger for a single Gemini CLI transaction."""
-
-    def __init__(self, model_name: str, enabled: bool = True):
-        self.enabled = enabled
-        if not self.enabled:
-            return
-
-        timestamp = datetime.now().strftime("%Y%m%d_%H%M%S_%f")
-        request_id = str(uuid.uuid4())
-        # Sanitize model name for directory
-        safe_model_name = model_name.replace("/", "_").replace(":", "_")
-        self.log_dir = (
-            _get_gemini_cli_logs_dir() / f"{timestamp}_{safe_model_name}_{request_id}"
-        )
-        try:
-            self.log_dir.mkdir(parents=True, exist_ok=True)
-        except Exception as e:
-            lib_logger.error(f"Failed to create Gemini CLI log directory: {e}")
-            self.enabled = False
-
-    def log_request(self, payload: Dict[str, Any]):
-        """Logs the request payload sent to Gemini."""
-        if not self.enabled:
-            return
-        try:
-            with open(
-                self.log_dir / "request_payload.json", "w", encoding="utf-8"
-            ) as f:
-                json.dump(payload, f, indent=2, ensure_ascii=False)
-        except Exception as e:
-            lib_logger.error(f"_GeminiCliFileLogger: Failed to write request: {e}")
-
-    def log_response_chunk(self, chunk: str):
-        """Logs a raw chunk from the Gemini response stream."""
-        if not self.enabled:
-            return
-        try:
-            with open(self.log_dir / "response_stream.log", "a", encoding="utf-8") as f:
-                f.write(chunk + "\n")
-        except Exception as e:
-            lib_logger.error(
-                f"_GeminiCliFileLogger: Failed to write response chunk: {e}"
-            )
-
-    def log_error(self, error_message: str):
-        """Logs an error message."""
-        if not self.enabled:
-            return
-        try:
-            with open(self.log_dir / "error.log", "a", encoding="utf-8") as f:
-                f.write(f"[{datetime.utcnow().isoformat()}] {error_message}\n")
-        except Exception as e:
-            lib_logger.error(f"_GeminiCliFileLogger: Failed to write error: {e}")
-
-    def log_final_response(self, response_data: Dict[str, Any]):
-        """Logs the final, reassembled response."""
-        if not self.enabled:
-            return
-        try:
-            with open(self.log_dir / "final_response.json", "w", encoding="utf-8") as f:
-                json.dump(response_data, f, indent=2, ensure_ascii=False)
-        except Exception as e:
-            lib_logger.error(
-                f"_GeminiCliFileLogger: Failed to write final response: {e}"
-            )
+# NOTE: _GeminiCliFileLogger has been moved to utilities.gemini_file_logger
+# and is imported as GeminiCliFileLogger
 
 
-CODE_ASSIST_ENDPOINT = "https://cloudcode-pa.googleapis.com/v1internal"
-
-HARDCODED_MODELS = [
+AVAILABLE_MODELS = [
     "gemini-2.5-pro",
     "gemini-2.5-flash",
     "gemini-2.5-flash-lite",
@@ -175,191 +115,19 @@ When in doubt, RE-READ THE SCHEMA before making the call.
 </CRITICAL_TOOL_USAGE_INSTRUCTIONS>
 """
 
-# Gemini finish reason mapping
-FINISH_REASON_MAP = {
-    "STOP": "stop",
-    "MAX_TOKENS": "length",
-    "SAFETY": "content_filter",
-    "RECITATION": "content_filter",
-    "OTHER": "stop",
-}
+# NOTE: FINISH_REASON_MAP has been moved to utilities.gemini_shared_utils
+
+# NOTE: _recursively_parse_json_strings, _inline_schema_refs, _env_bool, _env_int
+# have been moved to utilities.gemini_shared_utils and are imported at top of file
 
 
-def _recursively_parse_json_strings(
-    obj: Any,
-    schema: Optional[Dict[str, Any]] = None,
-    parse_json_objects: bool = False,
-) -> Any:
-    """
-    Recursively parse JSON strings in nested data structures.
-
-    Gemini sometimes returns tool arguments with JSON-stringified values:
-    {"files": "[{...}]"} instead of {"files": [{...}]}.
-
-    Args:
-        obj: The object to process
-        schema: Optional JSON schema for the current level (used for schema-aware parsing)
-        parse_json_objects: If False (default), don't parse JSON-looking strings into objects.
-                           This prevents corrupting string content like write tool's "content" field.
-                           If True, parse strings that look like JSON objects/arrays.
-
-    Additionally handles:
-    - Malformed double-encoded JSON (extra trailing '}' or ']') - only when parse_json_objects=True
-    - Escaped string content (\n, \t, etc.) - always processed
-    """
-    if isinstance(obj, dict):
-        # Get properties schema for looking up field types
-        properties_schema = schema.get("properties", {}) if schema else {}
-        return {
-            k: _recursively_parse_json_strings(
-                v,
-                properties_schema.get(k),
-                parse_json_objects,
-            )
-            for k, v in obj.items()
-        }
-    elif isinstance(obj, list):
-        # Get items schema for array elements
-        items_schema = schema.get("items") if schema else None
-        return [
-            _recursively_parse_json_strings(item, items_schema, parse_json_objects)
-            for item in obj
-        ]
-    elif isinstance(obj, str):
-        stripped = obj.strip()
-
-        # Check if string contains control character escape sequences that need unescaping
-        # This handles cases where diff content has literal \n or \t instead of actual newlines/tabs
-        #
-        # IMPORTANT: We intentionally do NOT unescape strings containing \" or \\
-        # because these are typically intentional escapes in code/config content
-        # (e.g., JSON embedded in YAML: BOT_NAMES_JSON: '["mirrobot", ...]')
-        # Unescaping these would corrupt the content and cause issues like
-        # oldString and newString becoming identical when they should differ.
-        has_control_char_escapes = "\\n" in obj or "\\t" in obj
-        has_intentional_escapes = '\\"' in obj or "\\\\" in obj
-
-        if has_control_char_escapes and not has_intentional_escapes:
-            try:
-                # Use json.loads with quotes to properly unescape the string
-                # This converts \n -> newline, \t -> tab
-                unescaped = json.loads(f'"{obj}"')
-                # Log the fix with a snippet for debugging
-                snippet = obj[:80] + "..." if len(obj) > 80 else obj
-                lib_logger.debug(
-                    f"[GeminiCli] Unescaped control chars in string: "
-                    f"{len(obj) - len(unescaped)} chars changed. Snippet: {snippet!r}"
-                )
-                return unescaped
-            except (json.JSONDecodeError, ValueError):
-                # If unescaping fails, continue with original processing
-                pass
-
-        # Only parse JSON strings if explicitly enabled
-        if not parse_json_objects:
-            return obj
-
-        # Schema-aware parsing: only parse if schema expects object/array, not string
-        if schema:
-            schema_type = schema.get("type")
-            if schema_type == "string":
-                # Schema says this should be a string - don't parse it
-                return obj
-            # Only parse if schema expects object or array
-            if schema_type not in ("object", "array", None):
-                return obj
-
-        # Check if it looks like JSON (starts with { or [)
-        if stripped and stripped[0] in ("{", "["):
-            # Try standard parsing first
-            if (stripped.startswith("{") and stripped.endswith("}")) or (
-                stripped.startswith("[") and stripped.endswith("]")
-            ):
-                try:
-                    parsed = json.loads(obj)
-                    return _recursively_parse_json_strings(
-                        parsed, schema, parse_json_objects
-                    )
-                except (json.JSONDecodeError, ValueError):
-                    pass
-
-            # Handle malformed JSON: array that doesn't end with ]
-            # e.g., '[{"path": "..."}]}' instead of '[{"path": "..."}]'
-            if stripped.startswith("[") and not stripped.endswith("]"):
-                try:
-                    # Find the last ] and truncate there
-                    last_bracket = stripped.rfind("]")
-                    if last_bracket > 0:
-                        cleaned = stripped[: last_bracket + 1]
-                        parsed = json.loads(cleaned)
-                        lib_logger.warning(
-                            f"[GeminiCli] Auto-corrected malformed JSON string: "
-                            f"truncated {len(stripped) - len(cleaned)} extra chars"
-                        )
-                        return _recursively_parse_json_strings(
-                            parsed, schema, parse_json_objects
-                        )
-                except (json.JSONDecodeError, ValueError):
-                    pass
-
-            # Handle malformed JSON: object that doesn't end with }
-            if stripped.startswith("{") and not stripped.endswith("}"):
-                try:
-                    # Find the last } and truncate there
-                    last_brace = stripped.rfind("}")
-                    if last_brace > 0:
-                        cleaned = stripped[: last_brace + 1]
-                        parsed = json.loads(cleaned)
-                        lib_logger.warning(
-                            f"[GeminiCli] Auto-corrected malformed JSON string: "
-                            f"truncated {len(stripped) - len(cleaned)} extra chars"
-                        )
-                        return _recursively_parse_json_strings(
-                            parsed, schema, parse_json_objects
-                        )
-                except (json.JSONDecodeError, ValueError):
-                    pass
-    return obj
-
-
-def _inline_schema_refs(schema: Dict[str, Any]) -> Dict[str, Any]:
-    """Inline local $ref definitions before sanitization."""
-    if not isinstance(schema, dict):
-        return schema
-
-    defs = schema.get("$defs", schema.get("definitions", {}))
-    if not defs:
-        return schema
-
-    def resolve(node, seen=()):
-        if not isinstance(node, dict):
-            return [resolve(x, seen) for x in node] if isinstance(node, list) else node
-        if "$ref" in node:
-            ref = node["$ref"]
-            if ref in seen:  # Circular - drop it
-                return {k: resolve(v, seen) for k, v in node.items() if k != "$ref"}
-            for prefix in ("#/$defs/", "#/definitions/"):
-                if isinstance(ref, str) and ref.startswith(prefix):
-                    name = ref[len(prefix) :]
-                    if name in defs:
-                        return resolve(copy.deepcopy(defs[name]), seen + (ref,))
-            return {k: resolve(v, seen) for k, v in node.items() if k != "$ref"}
-        return {k: resolve(v, seen) for k, v in node.items()}
-
-    return resolve(schema)
-
-
-def _env_bool(key: str, default: bool = False) -> bool:
-    """Get boolean from environment variable."""
-    return os.getenv(key, str(default).lower()).lower() in ("true", "1", "yes")
-
-
-def _env_int(key: str, default: int) -> int:
-    """Get integer from environment variable."""
-    return int(os.getenv(key, str(default)))
-
-
-class GeminiCliProvider(GeminiAuthBase, ProviderInterface):
+class GeminiCliProvider(
+    GeminiAuthBase,
+    GeminiCliQuotaTracker,
+    GeminiToolHandler,
+    GeminiCredentialManager,
+    ProviderInterface,
+):
     skip_cost_calculation = True
 
     # Sequential mode - stick with one credential until it gets a 429, then switch
@@ -389,21 +157,41 @@ class GeminiCliProvider(GeminiAuthBase, ProviderInterface):
     # Default priority for tiers not in the mapping
     default_tier_priority: int = 10
 
-    # Gemini CLI uses default daily reset - no custom usage_reset_configs
-    # (Empty dict means inherited get_usage_reset_config returns None)
+    # Usage reset configs for Gemini CLI
+    # Verified 2026-01-07: 24-hour fixed window from first request for ALL tiers
+    # The reset time is set when the first request is made and does NOT roll forward
+    usage_reset_configs = {
+        "default": UsageResetConfigDef(
+            window_seconds=24 * 60 * 60,  # 24 hours
+            mode="per_model",
+            description="24-hour per-model window (all tiers)",
+            field_name="models",
+        ),
+    }
 
-    # No quota groups defined for Gemini CLI
-    # (Models don't share quotas)
+    # Model quota groups - models that share quota/cooldown timing
+    # Verified 2026-01-07 via quota verification tests
+    # Can be overridden via env: QUOTA_GROUPS_GEMINI_CLI_{GROUP}="model1,model2"
+    model_quota_groups: QuotaGroupMap = {
+        # Pro models share a quota pool (verified: gemini-2.5-pro and gemini-3-pro-preview)
+        "pro": ["gemini-2.5-pro", "gemini-3-pro-preview"],
+        # All 2.x Flash models share a quota pool (verified: 2.0 shares with 2.5)
+        # Note: contrary to PR #62 which claimed 2.0-flash was standalone
+        "25-flash": ["gemini-2.0-flash", "gemini-2.5-flash", "gemini-2.5-flash-lite"],
+        # Gemini 3 Flash is standalone (verified)
+        "3-flash": ["gemini-3-flash-preview"],
+    }
 
     # Priority-based concurrency multipliers
     # Same structure as Antigravity (by coincidence, tiers share naming)
     # Priority 1 (paid ultra): 5x concurrent requests
     # Priority 2 (standard paid): 3x concurrent requests
-    # Others: 1x (no sequential fallback, uses global default)
+    # Others: Use sequential fallback (2x) or balanced default (1x)
     default_priority_multipliers = {1: 5, 2: 3}
 
-    # No sequential fallback for Gemini CLI (uses balanced mode default)
-    # default_sequential_fallback_multiplier = 1  (inherited from ProviderInterface)
+    # For sequential mode, lower priority tiers still get 2x to maintain stickiness
+    # For balanced mode, this doesn't apply (falls back to 1x)
+    default_sequential_fallback_multiplier = 2
 
     @staticmethod
     def parse_quota_error(
@@ -561,9 +349,15 @@ class GeminiCliProvider(GeminiAuthBase, ProviderInterface):
         self.model_definitions = ModelDefinitions()
         # NOTE: project_id_cache and project_tier_cache are inherited from GeminiAuthBase
 
+        # Quota refresh interval (mirrors Antigravity pattern)
+        self._quota_refresh_interval = env_int("GEMINI_CLI_QUOTA_REFRESH_INTERVAL", 300)
+
+        # Track whether initial quota fetch has been done (for background job)
+        self._initial_quota_fetch_done = False
+
         # Gemini 3 configuration from environment
-        memory_ttl = _env_int("GEMINI_CLI_SIGNATURE_CACHE_TTL", 3600)
-        disk_ttl = _env_int("GEMINI_CLI_SIGNATURE_DISK_TTL", 86400)
+        memory_ttl = env_int("GEMINI_CLI_SIGNATURE_CACHE_TTL", 3600)
+        disk_ttl = env_int("GEMINI_CLI_SIGNATURE_DISK_TTL", 86400)
 
         # Initialize signature cache for Gemini 3 thoughtSignatures
         self._signature_cache = ProviderCache(
@@ -574,20 +368,20 @@ class GeminiCliProvider(GeminiAuthBase, ProviderInterface):
         )
 
         # Gemini 3 feature flags
-        self._preserve_signatures_in_client = _env_bool(
+        self._preserve_signatures_in_client = env_bool(
             "GEMINI_CLI_PRESERVE_THOUGHT_SIGNATURES", True
         )
-        self._enable_signature_cache = _env_bool(
+        self._enable_signature_cache = env_bool(
             "GEMINI_CLI_ENABLE_SIGNATURE_CACHE", True
         )
-        self._enable_gemini3_tool_fix = _env_bool("GEMINI_CLI_GEMINI3_TOOL_FIX", True)
-        self._gemini3_enforce_strict_schema = _env_bool(
+        self._enable_gemini3_tool_fix = env_bool("GEMINI_CLI_GEMINI3_TOOL_FIX", True)
+        self._gemini3_enforce_strict_schema = env_bool(
             "GEMINI_CLI_GEMINI3_STRICT_SCHEMA", True
         )
         # Toggle for JSON string parsing in tool call arguments
         # NOTE: This is possibly redundant - modern Gemini models may not need this fix.
         # Disabled by default. Enable if you see JSON-stringified values in tool args.
-        self._enable_json_string_parsing = _env_bool(
+        self._enable_json_string_parsing = env_bool(
             "GEMINI_CLI_ENABLE_JSON_STRING_PARSING", False
         )
 
@@ -609,6 +403,10 @@ class GeminiCliProvider(GeminiAuthBase, ProviderInterface):
             f"gemini3_strict_schema={self._gemini3_enforce_strict_schema}"
         )
 
+        # Quota tracking instance variables (required by GeminiCliQuotaTracker mixin)
+        self._learned_costs: Dict[str, Dict[str, float]] = {}
+        self._learned_costs_loaded: bool = False
+
     # =========================================================================
     # CREDENTIAL TIER LOOKUP (Provider-specific - uses cache)
     # =========================================================================
@@ -616,66 +414,19 @@ class GeminiCliProvider(GeminiAuthBase, ProviderInterface):
     # NOTE: get_credential_priority() is now inherited from ProviderInterface.
     # It uses get_credential_tier_name() to get the tier and resolve priority
     # from the tier_priorities class attribute.
+    #
+    # NOTE: _load_tier_from_file(), get_credential_tier_name(), initialize_credentials(),
+    # _load_persisted_tiers(), get_background_job_config(), and run_background_job()
+    # are now inherited from GeminiCredentialManager mixin.
     # =========================================================================
 
-    def _load_tier_from_file(self, credential_path: str) -> Optional[str]:
-        """
-        Load tier from credential file's _proxy_metadata and cache it.
+    # NOTE: _load_tier_from_file() is inherited from GeminiCredentialManager
 
-        This is used as a fallback when the tier isn't in the memory cache,
-        typically on first access before initialize_credentials() has run.
-
-        Args:
-            credential_path: Path to the credential file
-
-        Returns:
-            Tier string if found, None otherwise
-        """
-        # Skip env:// paths (environment-based credentials)
-        if self._parse_env_credential_path(credential_path) is not None:
-            return None
-
-        try:
-            with open(credential_path, "r") as f:
-                creds = json.load(f)
-
-            metadata = creds.get("_proxy_metadata", {})
-            tier = metadata.get("tier")
-            project_id = metadata.get("project_id")
-
-            if tier:
-                self.project_tier_cache[credential_path] = tier
-                lib_logger.debug(
-                    f"Lazy-loaded tier '{tier}' for credential: {Path(credential_path).name}"
-                )
-
-            if project_id and credential_path not in self.project_id_cache:
-                self.project_id_cache[credential_path] = project_id
-
-            return tier
-        except (FileNotFoundError, json.JSONDecodeError, KeyError) as e:
-            lib_logger.debug(f"Could not lazy-load tier from {credential_path}: {e}")
-            return None
-
-    def get_credential_tier_name(self, credential: str) -> Optional[str]:
-        """
-        Returns the human-readable tier name for a credential.
-
-        Args:
-            credential: The credential path
-
-        Returns:
-            Tier name string (e.g., "free-tier") or None if unknown
-        """
-        tier = self.project_tier_cache.get(credential)
-        if not tier:
-            tier = self._load_tier_from_file(credential)
-        return tier
+    # NOTE: get_credential_tier_name() is inherited from GeminiCredentialManager
 
     def get_model_tier_requirement(self, model: str) -> Optional[int]:
         """
         Returns the minimum priority tier required for a model.
-        Gemini 3 requires paid tier (priority 1).
 
         Args:
             model: The model name (with or without provider prefix)
@@ -683,118 +434,21 @@ class GeminiCliProvider(GeminiAuthBase, ProviderInterface):
         Returns:
             Minimum required priority level or None if no restrictions
         """
-        model_name = model.split("/")[-1].replace(":thinking", "")
+        # No model-specific priority restrictions
+        # (Gemini 3 is now public and available to all tiers)
+        return None
 
-        # Gemini 3 requires paid tier
-        if model_name.startswith("gemini-3-"):
-            return 2  # Only priority 2 (paid) credentials
+    # NOTE: initialize_credentials() is inherited from GeminiCredentialManager
 
-        return None  # All other models have no restrictions
+    # NOTE: _load_persisted_tiers() is inherited from GeminiCredentialManager
 
-    async def initialize_credentials(self, credential_paths: List[str]) -> None:
-        """
-        Load persisted tier information from credential files at startup.
+    # =========================================================================
+    # BACKGROUND JOB INTERFACE (Quota Baseline Refresh)
+    # =========================================================================
 
-        This ensures all credential priorities are known before any API calls,
-        preventing unknown credentials from getting priority 999.
+    # NOTE: get_background_job_config() is inherited from GeminiCredentialManager
 
-        For credentials without persisted tier info (new or corrupted), performs
-        full discovery to ensure proper prioritization in sequential rotation mode.
-        """
-        # Step 1: Load persisted tiers from files
-        await self._load_persisted_tiers(credential_paths)
-
-        # Step 2: Identify credentials still missing tier info
-        credentials_needing_discovery = [
-            path
-            for path in credential_paths
-            if path not in self.project_tier_cache
-            and self._parse_env_credential_path(path) is None  # Skip env:// paths
-        ]
-
-        if not credentials_needing_discovery:
-            return  # All credentials have tier info
-
-        lib_logger.info(
-            f"GeminiCli: Discovering tier info for {len(credentials_needing_discovery)} credential(s)..."
-        )
-
-        # Step 3: Perform discovery for each missing credential (sequential to avoid rate limits)
-        for credential_path in credentials_needing_discovery:
-            try:
-                auth_header = await self.get_auth_header(credential_path)
-                access_token = auth_header["Authorization"].split(" ")[1]
-                await self._discover_project_id(
-                    credential_path, access_token, litellm_params={}
-                )
-                discovered_tier = self.project_tier_cache.get(
-                    credential_path, "unknown"
-                )
-                lib_logger.debug(
-                    f"Discovered tier '{discovered_tier}' for {Path(credential_path).name}"
-                )
-            except Exception as e:
-                lib_logger.warning(
-                    f"Failed to discover tier for {Path(credential_path).name}: {e}. "
-                    f"Credential will use default priority."
-                )
-
-    async def _load_persisted_tiers(
-        self, credential_paths: List[str]
-    ) -> Dict[str, str]:
-        """
-        Load persisted tier information from credential files into memory cache.
-
-        Args:
-            credential_paths: List of credential file paths
-
-        Returns:
-            Dict mapping credential path to tier name for logging purposes
-        """
-        loaded = {}
-        for path in credential_paths:
-            # Skip env:// paths (environment-based credentials)
-            if self._parse_env_credential_path(path) is not None:
-                continue
-
-            # Skip if already in cache
-            if path in self.project_tier_cache:
-                continue
-
-            try:
-                with open(path, "r") as f:
-                    creds = json.load(f)
-
-                metadata = creds.get("_proxy_metadata", {})
-                tier = metadata.get("tier")
-                project_id = metadata.get("project_id")
-
-                if tier:
-                    self.project_tier_cache[path] = tier
-                    loaded[path] = tier
-                    lib_logger.debug(
-                        f"Loaded persisted tier '{tier}' for credential: {Path(path).name}"
-                    )
-
-                if project_id:
-                    self.project_id_cache[path] = project_id
-
-            except (FileNotFoundError, json.JSONDecodeError, KeyError) as e:
-                lib_logger.debug(f"Could not load persisted tier from {path}: {e}")
-
-        if loaded:
-            # Log summary at debug level
-            tier_counts: Dict[str, int] = {}
-            for tier in loaded.values():
-                tier_counts[tier] = tier_counts.get(tier, 0) + 1
-            lib_logger.debug(
-                f"GeminiCli: Loaded {len(loaded)} credential tiers from disk: "
-                + ", ".join(
-                    f"{tier}={count}" for tier, count in sorted(tier_counts.items())
-                )
-            )
-
-        return loaded
+    # NOTE: run_background_job() is inherited from GeminiCredentialManager
 
     # NOTE: _post_auth_discovery() is inherited from GeminiAuthBase
 
@@ -807,17 +461,19 @@ class GeminiCliProvider(GeminiAuthBase, ProviderInterface):
         model_name = model.split("/")[-1].replace(":thinking", "")
         return model_name.startswith("gemini-3-")
 
-    def _strip_gemini3_prefix(self, name: str) -> str:
+    def _get_available_models(self) -> List[str]:
         """
-        Strip the Gemini 3 namespace prefix from a tool name.
+        Get list of user-facing model names available via this provider.
 
-        Also reverses any tool renames that were applied to avoid Gemini conflicts.
+        Used by quota tracker to filter which models to store baselines for.
+        Only models in this list will have quota baselines tracked.
+
+        Returns:
+            List of user-facing model names
         """
-        if name and name.startswith(self._gemini3_tool_prefix):
-            stripped = name[len(self._gemini3_tool_prefix) :]
-            # Reverse any renames
-            return GEMINI3_TOOL_RENAMES_REVERSE.get(stripped, stripped)
-        return name
+        return AVAILABLE_MODELS
+
+    # NOTE: _strip_gemini3_prefix() is inherited from GeminiToolHandler
 
     # NOTE: _discover_project_id() and _persist_project_metadata() are inherited from GeminiAuthBase
 
@@ -1072,228 +728,19 @@ class GeminiCliProvider(GeminiAuthBase, ProviderInterface):
 
         return system_instruction, gemini_contents
 
-    def _fix_tool_response_grouping(
-        self, contents: List[Dict[str, Any]]
-    ) -> List[Dict[str, Any]]:
-        """
-        Group function calls with their responses for Gemini CLI compatibility.
-
-        Converts linear format (call, response, call, response)
-        to grouped format (model with calls, user with all responses).
-
-        IMPORTANT: Preserves ID-based pairing to prevent mismatches.
-        When IDs don't match, attempts recovery by:
-        1. Matching by function name first
-        2. Matching by order if names don't match
-        3. Inserting placeholder responses if responses are missing
-        4. Inserting responses at the CORRECT position (after their corresponding call)
-        """
-        new_contents = []
-        # Each pending group tracks:
-        # - ids: expected response IDs
-        # - func_names: expected function names (for orphan matching)
-        # - insert_after_idx: position in new_contents where model message was added
-        pending_groups = []
-        collected_responses = {}  # Dict mapping ID -> response_part
-
-        for content in contents:
-            role = content.get("role")
-            parts = content.get("parts", [])
-
-            response_parts = [p for p in parts if "functionResponse" in p]
-
-            if response_parts:
-                # Collect responses by ID (ignore duplicates - keep first occurrence)
-                for resp in response_parts:
-                    resp_id = resp.get("functionResponse", {}).get("id", "")
-                    if resp_id:
-                        if resp_id in collected_responses:
-                            lib_logger.warning(
-                                f"[Grouping] Duplicate response ID detected: {resp_id}. "
-                                f"Ignoring duplicate - this may indicate malformed conversation history."
-                            )
-                            continue
-                        collected_responses[resp_id] = resp
-
-                # Try to satisfy pending groups (newest first)
-                for i in range(len(pending_groups) - 1, -1, -1):
-                    group = pending_groups[i]
-                    group_ids = group["ids"]
-
-                    # Check if we have ALL responses for this group
-                    if all(gid in collected_responses for gid in group_ids):
-                        # Extract responses in the same order as the function calls
-                        group_responses = [
-                            collected_responses.pop(gid) for gid in group_ids
-                        ]
-                        new_contents.append({"parts": group_responses, "role": "user"})
-                        pending_groups.pop(i)
-                        break
-                continue
-
-            if role == "model":
-                func_calls = [p for p in parts if "functionCall" in p]
-                new_contents.append(content)
-                if func_calls:
-                    call_ids = [
-                        fc.get("functionCall", {}).get("id", "") for fc in func_calls
-                    ]
-                    call_ids = [cid for cid in call_ids if cid]  # Filter empty IDs
-
-                    # Also extract function names for orphan matching
-                    func_names = [
-                        fc.get("functionCall", {}).get("name", "") for fc in func_calls
-                    ]
-
-                    if call_ids:
-                        pending_groups.append(
-                            {
-                                "ids": call_ids,
-                                "func_names": func_names,
-                                "insert_after_idx": len(new_contents) - 1,
-                            }
-                        )
-            else:
-                new_contents.append(content)
-
-        # Handle remaining groups (shouldn't happen in well-formed conversations)
-        # Attempt recovery by matching orphans to unsatisfied calls
-        # Process in REVERSE order of insert_after_idx so insertions don't shift indices
-        pending_groups.sort(key=lambda g: g["insert_after_idx"], reverse=True)
-
-        for group in pending_groups:
-            group_ids = group["ids"]
-            group_func_names = group.get("func_names", [])
-            insert_idx = group["insert_after_idx"] + 1
-            group_responses = []
-
-            lib_logger.debug(
-                f"[Grouping Recovery] Processing unsatisfied group: "
-                f"ids={group_ids}, names={group_func_names}, insert_at={insert_idx}"
-            )
-
-            for i, expected_id in enumerate(group_ids):
-                expected_name = group_func_names[i] if i < len(group_func_names) else ""
-
-                if expected_id in collected_responses:
-                    # Direct ID match
-                    group_responses.append(collected_responses.pop(expected_id))
-                    lib_logger.debug(
-                        f"[Grouping Recovery] Direct ID match for '{expected_id}'"
-                    )
-                elif collected_responses:
-                    # Try to find orphan with matching function name first
-                    matched_orphan_id = None
-
-                    # First pass: match by function name
-                    for orphan_id, orphan_resp in collected_responses.items():
-                        orphan_name = orphan_resp.get("functionResponse", {}).get(
-                            "name", ""
-                        )
-                        # Match if names are equal
-                        if orphan_name == expected_name:
-                            matched_orphan_id = orphan_id
-                            lib_logger.debug(
-                                f"[Grouping Recovery] Matched orphan '{orphan_id}' by name '{orphan_name}'"
-                            )
-                            break
-
-                    # Second pass: if no name match, try "unknown_function" orphans
-                    if not matched_orphan_id:
-                        for orphan_id, orphan_resp in collected_responses.items():
-                            orphan_name = orphan_resp.get("functionResponse", {}).get(
-                                "name", ""
-                            )
-                            if orphan_name == "unknown_function":
-                                matched_orphan_id = orphan_id
-                                lib_logger.debug(
-                                    f"[Grouping Recovery] Matched unknown_function orphan '{orphan_id}' "
-                                    f"to expected '{expected_name}'"
-                                )
-                                break
-
-                    # Third pass: if still no match, take first available (order-based)
-                    if not matched_orphan_id:
-                        matched_orphan_id = next(iter(collected_responses))
-                        lib_logger.debug(
-                            f"[Grouping Recovery] No name match, using first available orphan '{matched_orphan_id}'"
-                        )
-
-                    if matched_orphan_id:
-                        orphan_resp = collected_responses.pop(matched_orphan_id)
-
-                        # Fix the ID in the response to match the call
-                        old_id = orphan_resp["functionResponse"].get("id", "")
-                        orphan_resp["functionResponse"]["id"] = expected_id
-
-                        # Fix the name if it was "unknown_function"
-                        if (
-                            orphan_resp["functionResponse"].get("name")
-                            == "unknown_function"
-                            and expected_name
-                        ):
-                            orphan_resp["functionResponse"]["name"] = expected_name
-                            lib_logger.info(
-                                f"[Grouping Recovery] Fixed function name from 'unknown_function' to '{expected_name}'"
-                            )
-
-                        lib_logger.warning(
-                            f"[Grouping] Auto-repaired ID mismatch: mapped response '{old_id}' "
-                            f"to call '{expected_id}' (function: {expected_name})"
-                        )
-                        group_responses.append(orphan_resp)
-                else:
-                    # No responses available - create placeholder
-                    placeholder_resp = {
-                        "functionResponse": {
-                            "name": expected_name or "unknown_function",
-                            "response": {
-                                "result": {
-                                    "error": "Tool response was lost during context processing. "
-                                    "This is a recovered placeholder.",
-                                    "recovered": True,
-                                }
-                            },
-                            "id": expected_id,
-                        }
-                    }
-                    lib_logger.warning(
-                        f"[Grouping Recovery] Created placeholder response for missing tool: "
-                        f"id='{expected_id}', name='{expected_name}'"
-                    )
-                    group_responses.append(placeholder_resp)
-
-            if group_responses:
-                # Insert at the correct position (right after the model message with the calls)
-                new_contents.insert(
-                    insert_idx, {"parts": group_responses, "role": "user"}
-                )
-                lib_logger.info(
-                    f"[Grouping Recovery] Inserted {len(group_responses)} responses at position {insert_idx} "
-                    f"(expected {len(group_ids)})"
-                )
-
-        # Warn about unmatched responses
-        if collected_responses:
-            lib_logger.warning(
-                f"[Grouping] {len(collected_responses)} unmatched responses remaining: "
-                f"ids={list(collected_responses.keys())}"
-            )
-
-        return new_contents
+    # NOTE: _fix_tool_response_grouping() is inherited from GeminiToolHandler mixin
 
     def _handle_reasoning_parameters(
         self, payload: Dict[str, Any], model: str
     ) -> Optional[Dict[str, Any]]:
         """
-        Map reasoning_effort to thinking configuration.
+        Map reasoning_effort to thinking configuration for Gemini models.
 
         - Gemini 2.5: thinkingBudget (integer tokens)
         - Gemini 3 Pro: thinkingLevel (string: "low"/"high")
         - Gemini 3 Flash: thinkingLevel (string: "minimal"/"low"/"medium"/"high")
         """
-        custom_reasoning_budget = payload.get("custom_reasoning_budget", False)
-        reasoning_effort = payload.get("reasoning_effort")
+        reasoning_effort = payload.pop("reasoning_effort", None)
 
         if "thinkingConfig" in payload.get("generationConfig", {}):
             return None
@@ -1302,68 +749,84 @@ class GeminiCliProvider(GeminiAuthBase, ProviderInterface):
         is_gemini_3 = self._is_gemini_3(model)
         is_gemini_3_flash = "gemini-3-flash" in model
 
-        # Only apply reasoning logic to supported models
         if not (is_gemini_25 or is_gemini_3):
-            payload.pop("reasoning_effort", None)
-            payload.pop("custom_reasoning_budget", None)
             return None
 
-        # Gemini 3 Flash: Supports minimal/low/medium/high thinkingLevel
-        if is_gemini_3_flash:
-            # Clean up the original payload
-            payload.pop("reasoning_effort", None)
-            payload.pop("custom_reasoning_budget", None)
+        # Normalize and validate upfront
+        if reasoning_effort is None:
+            effort = "auto"
+        elif isinstance(reasoning_effort, str):
+            effort = reasoning_effort.strip().lower() or "auto"
+        else:
+            lib_logger.warning(
+                f"[GeminiCLI] Invalid reasoning_effort type: {type(reasoning_effort).__name__}, using auto"
+            )
+            effort = "auto"
 
-            if reasoning_effort == "disable":
-                # "minimal" matches "no thinking" for most queries
+        valid_efforts = {
+            "auto",
+            "disable",
+            "off",
+            "none",
+            "minimal",
+            "low",
+            "low_medium",
+            "medium",
+            "medium_high",
+            "high",
+        }
+        if effort not in valid_efforts:
+            lib_logger.warning(
+                f"[GeminiCLI] Unknown reasoning_effort: '{reasoning_effort}', using auto"
+            )
+            effort = "auto"
+
+        # Gemini 3 Flash: minimal/low/medium/high
+        if is_gemini_3_flash:
+            if effort in ("disable", "off", "none"):
                 return {"thinkingLevel": "minimal", "include_thoughts": True}
-            elif reasoning_effort == "low":
+            if effort in ("minimal", "low"):
                 return {"thinkingLevel": "low", "include_thoughts": True}
-            elif reasoning_effort == "medium":
+            if effort in ("low_medium", "medium"):
                 return {"thinkingLevel": "medium", "include_thoughts": True}
-            # Default to high for Flash
+            # auto, medium_high, high → high
             return {"thinkingLevel": "high", "include_thoughts": True}
 
-        # Gemini 3 Pro: Only supports low/high thinkingLevel
+        # Gemini 3 Pro: only low/high
         if is_gemini_3:
-            # Clean up the original payload
-            payload.pop("reasoning_effort", None)
-            payload.pop("custom_reasoning_budget", None)
-
-            if reasoning_effort == "low":
+            if effort in ("disable", "off", "none", "minimal", "low", "low_medium"):
                 return {"thinkingLevel": "low", "include_thoughts": True}
-            # medium maps to high for Pro (not supported)
+            # auto, medium, medium_high, high → high
             return {"thinkingLevel": "high", "include_thoughts": True}
 
         # Gemini 2.5: Integer thinkingBudget
-        if not reasoning_effort:
-            # Clean up the original payload
-            payload.pop("reasoning_effort", None)
-            payload.pop("custom_reasoning_budget", None)
+        if effort in ("disable", "off", "none"):
+            return {"thinkingBudget": 0, "include_thoughts": False}
+
+        if effort == "auto":
             return {"thinkingBudget": -1, "include_thoughts": True}
 
-        # If reasoning_effort is provided, calculate the budget
-        budget = -1  # Default for 'auto' or invalid values
-        if "gemini-2.5-pro" in model:
-            budgets = {"low": 8192, "medium": 16384, "high": 32768}
-        elif "gemini-2.5-flash" in model:
-            budgets = {"low": 6144, "medium": 12288, "high": 24576}
+        # Model-specific budgets
+        if "gemini-2.5-flash" in model:
+            budgets = {
+                "minimal": 3072,
+                "low": 6144,
+                "low_medium": 9216,
+                "medium": 12288,
+                "medium_high": 18432,
+                "high": 24576,
+            }
         else:
-            # Fallback for other gemini-2.5 models
-            budgets = {"low": 1024, "medium": 2048, "high": 4096}
+            budgets = {
+                "minimal": 4096,
+                "low": 8192,
+                "low_medium": 12288,
+                "medium": 16384,
+                "medium_high": 24576,
+                "high": 32768,
+            }
 
-        budget = budgets.get(reasoning_effort, -1)
-        if reasoning_effort == "disable":
-            budget = 0
-
-        if not custom_reasoning_budget:
-            budget = budget // 4
-
-        # Clean up the original payload
-        payload.pop("reasoning_effort", None)
-        payload.pop("custom_reasoning_budget", None)
-
-        return {"thinkingBudget": budget, "include_thoughts": True}
+        return {"thinkingBudget": budgets[effort], "include_thoughts": True}
 
     def _convert_chunk_to_openai(
         self,
@@ -1424,7 +887,7 @@ class GeminiCliProvider(GeminiAuthBase, ProviderInterface):
                 # NOTE: This is very possibly redundant
                 raw_args = function_call.get("args", {})
                 if self._enable_json_string_parsing:
-                    tool_args = _recursively_parse_json_strings(raw_args)
+                    tool_args = recursively_parse_json_strings(raw_args)
                 else:
                     tool_args = raw_args
 
@@ -1693,54 +1156,7 @@ class GeminiCliProvider(GeminiAuthBase, ProviderInterface):
 
         return schema
 
-    def _enforce_strict_schema(self, schema: Any) -> Any:
-        """
-        Enforce strict JSON schema for Gemini 3 to prevent hallucinated parameters.
-
-        Adds 'additionalProperties: false' to object schemas with 'properties',
-        which tells the model it CANNOT add properties not in the schema.
-
-        IMPORTANT: Preserves 'additionalProperties: true' (or {}) when explicitly
-        set in the original schema. This is critical for "freeform" parameter objects
-        like batch/multi_tool's nested parameters which need to accept arbitrary
-        tool parameters that aren't pre-defined in the schema.
-        """
-        if not isinstance(schema, dict):
-            return schema
-
-        result = {}
-        preserved_additional_props = None
-
-        for key, value in schema.items():
-            # Preserve additionalProperties as-is if it's truthy
-            # This is critical for "freeform" parameter objects like batch's
-            # nested parameters which need to accept arbitrary tool parameters
-            if key == "additionalProperties":
-                if value is not False:
-                    # Preserve the original value (true, {}, {"type": "string"}, etc.)
-                    preserved_additional_props = value
-                continue
-            if isinstance(value, dict):
-                result[key] = self._enforce_strict_schema(value)
-            elif isinstance(value, list):
-                result[key] = [
-                    self._enforce_strict_schema(item)
-                    if isinstance(item, dict)
-                    else item
-                    for item in value
-                ]
-            else:
-                result[key] = value
-
-        # Add additionalProperties: false to object schemas with properties,
-        # BUT only if we didn't preserve a value from the original schema
-        if result.get("type") == "object" and "properties" in result:
-            if preserved_additional_props is not None:
-                result["additionalProperties"] = preserved_additional_props
-            else:
-                result["additionalProperties"] = False
-
-        return result
+    # NOTE: _enforce_strict_schema() is inherited from GeminiToolHandler mixin
 
     def _transform_tool_schemas(
         self, tools: List[Dict[str, Any]], model: str = ""
@@ -1767,7 +1183,7 @@ class GeminiCliProvider(GeminiAuthBase, ProviderInterface):
                 # Gemini CLI expects 'parametersJsonSchema' instead of 'parameters'
                 if "parameters" in new_function:
                     # Inline $ref definitions first
-                    schema = _inline_schema_refs(new_function["parameters"])
+                    schema = inline_schema_refs(new_function["parameters"])
                     schema = self._gemini_cli_transform_schema(schema)
                     # Workaround: Gemini fails to emit functionCall for tools
                     # with empty properties {}. Inject a required confirmation param.
@@ -1817,96 +1233,18 @@ class GeminiCliProvider(GeminiAuthBase, ProviderInterface):
                         )
 
                     # Inject parameter signature into description
-                    new_function = self._inject_signature_into_description(new_function)
+                    new_function = self._inject_signature_into_description(
+                        new_function, self._gemini3_description_prompt
+                    )
 
                 transformed_declarations.append(new_function)
 
         return transformed_declarations
 
-    def _inject_signature_into_description(
-        self, func_decl: Dict[str, Any]
-    ) -> Dict[str, Any]:
-        """Inject parameter signatures into tool description for Gemini 3."""
-        schema = func_decl.get("parametersJsonSchema", {})
-        if not schema:
-            return func_decl
+    # NOTE: _inject_signature_into_description() is inherited from GeminiToolHandler mixin
+    # The inherited version requires passing the description_prompt parameter
 
-        required = schema.get("required", [])
-        properties = schema.get("properties", {})
-
-        if not properties:
-            return func_decl
-
-        param_list = []
-        for prop_name, prop_data in properties.items():
-            if not isinstance(prop_data, dict):
-                continue
-
-            type_hint = self._format_type_hint(prop_data)
-            is_required = prop_name in required
-            param_list.append(
-                f"{prop_name} ({type_hint}{', REQUIRED' if is_required else ''})"
-            )
-
-        if param_list:
-            sig_str = self._gemini3_description_prompt.replace(
-                "{params}", ", ".join(param_list)
-            )
-            func_decl["description"] = func_decl.get("description", "") + sig_str
-
-        return func_decl
-
-    def _format_type_hint(self, prop_data: Dict[str, Any], depth: int = 0) -> str:
-        """Format a detailed type hint for a property schema."""
-        type_hint = prop_data.get("type", "unknown")
-
-        # Handle enum values - show allowed options
-        if "enum" in prop_data:
-            enum_vals = prop_data["enum"]
-            if len(enum_vals) <= 5:
-                return f"string ENUM[{', '.join(repr(v) for v in enum_vals)}]"
-            return f"string ENUM[{len(enum_vals)} options]"
-
-        # Handle const values
-        if "const" in prop_data:
-            return f"string CONST={repr(prop_data['const'])}"
-
-        if type_hint == "array":
-            items = prop_data.get("items", {})
-            if isinstance(items, dict):
-                item_type = items.get("type", "unknown")
-                if item_type == "object":
-                    nested_props = items.get("properties", {})
-                    nested_req = items.get("required", [])
-                    if nested_props:
-                        nested_list = []
-                        for n, d in nested_props.items():
-                            if isinstance(d, dict):
-                                # Recursively format nested types (limit depth)
-                                if depth < 1:
-                                    t = self._format_type_hint(d, depth + 1)
-                                else:
-                                    t = d.get("type", "unknown")
-                                req = " REQUIRED" if n in nested_req else ""
-                                nested_list.append(f"{n}: {t}{req}")
-                        return f"ARRAY_OF_OBJECTS[{', '.join(nested_list)}]"
-                    return "ARRAY_OF_OBJECTS"
-                return f"ARRAY_OF_{item_type.upper()}"
-            return "ARRAY"
-
-        if type_hint == "object":
-            nested_props = prop_data.get("properties", {})
-            nested_req = prop_data.get("required", [])
-            if nested_props and depth < 1:
-                nested_list = []
-                for n, d in nested_props.items():
-                    if isinstance(d, dict):
-                        t = d.get("type", "unknown")
-                        req = " REQUIRED" if n in nested_req else ""
-                        nested_list.append(f"{n}: {t}{req}")
-                return f"object{{{', '.join(nested_list)}}}"
-
-        return type_hint
+    # NOTE: _format_type_hint() is inherited from GeminiToolHandler mixin
 
     def _inject_gemini3_system_instruction(
         self, request_payload: Dict[str, Any]
@@ -1935,53 +1273,14 @@ class GeminiCliProvider(GeminiAuthBase, ProviderInterface):
                 "parts": [{"text": self._gemini3_system_instruction}],
             }
 
-    def _translate_tool_choice(
-        self, tool_choice: Union[str, Dict[str, Any]], model: str = ""
-    ) -> Optional[Dict[str, Any]]:
-        """
-        Translates OpenAI's `tool_choice` to Gemini's `toolConfig`.
-        Handles Gemini 3 namespace prefixes for specific tool selection.
-        """
-        if not tool_choice:
-            return None
-
-        config = {}
-        mode = "AUTO"  # Default to auto
-        is_gemini_3 = self._is_gemini_3(model)
-
-        if isinstance(tool_choice, str):
-            if tool_choice == "auto":
-                mode = "AUTO"
-            elif tool_choice == "none":
-                mode = "NONE"
-            elif tool_choice == "required":
-                mode = "ANY"
-        elif isinstance(tool_choice, dict) and tool_choice.get("type") == "function":
-            function_name = tool_choice.get("function", {}).get("name")
-            if function_name:
-                # Add Gemini 3 prefix if needed (and rename problematic tools)
-                if is_gemini_3 and self._enable_gemini3_tool_fix:
-                    function_name = GEMINI3_TOOL_RENAMES.get(
-                        function_name, function_name
-                    )
-                    function_name = f"{self._gemini3_tool_prefix}{function_name}"
-
-                mode = "ANY"  # Force a call, but only to this function
-                config["functionCallingConfig"] = {
-                    "mode": mode,
-                    "allowedFunctionNames": [function_name],
-                }
-                return config
-
-        config["functionCallingConfig"] = {"mode": mode}
-        return config
+    # NOTE: _translate_tool_choice() is inherited from GeminiToolHandler mixin
 
     async def acompletion(
         self, client: httpx.AsyncClient, **kwargs
     ) -> Union[litellm.ModelResponse, AsyncGenerator[litellm.ModelResponse, None]]:
         model = kwargs["model"]
         credential_path = kwargs.pop("credential_identifier")
-        enable_request_logging = kwargs.pop("enable_request_logging", False)
+        transaction_context = kwargs.pop("transaction_context", None)
 
         # Get fallback models for rate limit handling
         fallback_models = self._cli_preview_fallback_order(model)
@@ -1998,24 +1297,11 @@ class GeminiCliProvider(GeminiAuthBase, ProviderInterface):
                     credential_path, access_token, kwargs.get("litellm_params", {})
                 )
 
-            # Log paid tier usage visibly on each request
-            credential_tier = self.project_tier_cache.get(credential_path)
-            if credential_tier and credential_tier not in [
-                "free-tier",
-                "legacy-tier",
-                "unknown",
-            ]:
-                lib_logger.info(
-                    f"[PAID TIER] Using Gemini '{credential_tier}' subscription for this request"
-                )
-
             # Handle :thinking suffix
             model_name = attempt_model.split("/")[-1].replace(":thinking", "")
 
-            # [NEW] Create a dedicated file logger for this request
-            file_logger = _GeminiCliFileLogger(
-                model_name=model_name, enabled=enable_request_logging
-            )
+            # Create provider logger from transaction context
+            file_logger = ProviderLogger(transaction_context)
 
             is_gemini_3 = self._is_gemini_3(model_name)
 
@@ -2384,8 +1670,8 @@ class GeminiCliProvider(GeminiAuthBase, ProviderInterface):
     async def get_models(self, credential: str, client: httpx.AsyncClient) -> List[str]:
         """
         Returns a merged list of Gemini CLI models from three sources:
-        1. Environment variable models (via GEMINI_CLI_MODELS) - ALWAYS included, take priority
-        2. Hardcoded models (fallback list) - added only if ID not in env vars
+        1. Environment variable models (via model definitions) - ALWAYS included, take priority
+        2. Available models (AVAILABLE_MODELS fallback list) - added only if ID not in env vars
         3. Dynamic discovery from Gemini API (if supported) - added only if ID not in env vars
 
         Environment variable models always win and are never deduplicated, even if they
@@ -2431,8 +1717,8 @@ class GeminiCliProvider(GeminiAuthBase, ProviderInterface):
                 f"Loaded {len(static_models)} static models for gemini_cli from environment variables"
             )
 
-        # Source 2: Add hardcoded models (only if ID not already in env vars)
-        for model_id in HARDCODED_MODELS:
+        # Source 2: Add available models (only if ID not already in env vars)
+        for model_id in AVAILABLE_MODELS:
             if model_id not in env_var_ids:
                 models.append(f"gemini_cli/{model_id}")
                 env_var_ids.add(model_id)
